+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 Changes to **.ENV Selector** can be found in this file.
 
+## - 0.2.0
+
+- Added regex to configuration
+- Added status bar item
+- Added Keyboard Shortcut
+- Added new logo
+- Added GIFs to the README.md.
+
 ## - 0.1.0
 
 - Initial release.